Stop trying to evaluate string annotations. Prepare for changes to PEP-649 annotations.
This removes a whole bunch of complicated code from `annotations.py`.

It may be dumb, but the check for ClassVar in string annotations is now: does the string contain 'ClassVar'.
Honestly this is probably good enough for 99% of cases.
The 1% can change their annotation to "ClassVar", or not "ClassVar" if it's a string annotation to avoid this behaviour as appropriate.